General cleanup before bed. Going to start designing the segment service after work... Current note: need to decide if we want to implment the DAO pattern or just let jpa do its thins.
I see huge advantages in both, and huge distavantages in both. Design docs following.

+++ src/main/java/com/clinkworks/timecard/service/EntryService.java
@@ -6,6 +6,7 @@ import java.util.Comparator;
 import java.util.List;
 
 import javax.persistence.EntityManager;
+import javax.persistence.EntityTransaction;
 import javax.persistence.TemporalType;
 
 import org.joda.time.DateTime;
@@ -61,6 +62,18 @@ public class EntryService{
 	@Transactional
 	public void deleteEntries(List<Entry> entries){
 		for(Entry entry : entries){
+			em.remove(entry);
+		}
+	}
+	
+	/**
+	 * deletes the entry with the specified id.
+	 * @param id
+	 */
+	@Transactional
+	public void deleteEntryById(Long id){
+		Entry entry = em.find(Entry.class, id);
+		if(entry != null){
 			em.remove(entry);
 		}
 	}
